Refactor code structure, simplify merge logic
