+++ src/commands/index.ts
@@ -1,0 +1,1 @@
+export * from './mergeThis';
